Manejo de los distintos tipos de modos que tiene una orden ademas de inicializacion de sucursal/store y datos iniciales refactorizado
