+++ src/main/java/com/sanisidro/restaurante/core/config/SecurityConfig.java
@@ -74,6 +74,7 @@ public class SecurityConfig {
                         .requestMatchers("/api/v1/reviews/**").hasAnyRole("CLIENT", "ADMIN")
                         .requestMatchers("/api/v1/orders/**").hasAnyRole("CLIENT", "WAITER", "CHEF", "CASHIER", "ADMIN")
                         .requestMatchers("/api/v1/inventory/**").hasAnyRole("ADMIN", "CHEF")
+                        .requestMatchers("/api/v1/stores/**").hasAnyRole("CLIENT", "ADMIN")
                         .anyRequest().authenticated()
                 )
                 .oauth2Login(oauth2 -> oauth2
